рабочий переброс файлов по своим местам
